pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alphapulldown"
description = "Pipeline allows massive screening using AlphaFold"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.8"
authors = [
    { name = "EMBL Hamburg", email = "alphapulldown@embl-hamburg.de" }
]
urls = { "Homepage" = "https://github.com/KosinskiLab/AlphaPulldown" }
classifiers = [
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: Apache Software License',
    'Operating System :: POSIX :: Linux',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
dependencies = [
    "absl-py >= 0.13.0",
    "dm-haiku",
    "dm-tree >= 0.1.6",
    "h5py >= 3.1.0",
    "matplotlib >= 3.3.3",
    "ml-collections >= 0.1.0",
    "pandas >= 1.5.3",
    "importlib-resources >= 6.1.0",
    "biopython >= 1.82",
    "nbformat >= 5.9.2",
    "py3Dmol == 2.0.4",
    "ipython == 8.16.1",
    "tqdm >= 4.66.1",
    "appdirs >= 1.4.4",
    "ml-dtypes",
    "setuptools >= 40.1.0",
    "jupyterlab >= 3.0",
    "ipywidgets",
    "pytest >= 6.0",
    "importlib-metadata >= 4.8.2, <5.0.0",
    "tensorflow-cpu >= 2.16.1",
    "chex >= 0.1.86",
    "immutabledict >= 2.0.0"
]
dynamic = ["version"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = [
    "alphapulldown",
    "alphapulldown.utils",
    "alphapulldown.folding_backend",
    "alphapulldown.analysis_pipeline",
    "alphapulldown.scripts",
    "alphafold",
    "alphafold.data",
    "alphafold.data.tools",
    "alphafold.common",
    "alphafold.relax",
    "alphafold.model",
    "alphafold.model.geometry",
    "alphafold.model.tf",
    "alphafold.notebooks",
    "colabfold",
    "analysis_pipeline",
    "af2plots",
    "unifold",
    "unifold.data",
    "unifold.modules",
    "unifold.msa",
    "unifold.losses",
    "unifold.symmetry"
]

[tool.setuptools.package-dir]
alphapulldown = "alphapulldown"
"alphapulldown.folding_backend" = "alphapulldown/folding_backend"
alphafold = "alphafold/alphafold"
colabfold = "ColabFold/colabfold"
analysis_pipeline = "alphapulldown/analysis_pipeline"
af2plots = "alphapulldown/analysis_pipeline/af2plots/af2plots"
unifold = "AlphaLink2/unifold"
"unifold.data" = "AlphaLink2/unifold/data"
"unifold.losses" = "AlphaLink2/unifold/losses"
"unifold.symmetry" = "AlphaLink2/unifold/symmetry"

[tool.setuptools.package-data]
"alphafold.common" = ["stereo_chemical_props.txt"]

[tool.setuptools.dynamic]
version = { attr = "alphapulldown.__version__" }

[project.scripts]
create_individual_features = "alphapulldown.scripts.entry_points:create_individual_features"
run_multimer_jobs = "alphapulldown.scripts.entry_points:run_multimer_jobs"
create_notebook = "alphapulldown.scripts.entry_points:create_notebook"
rename_colab_search_a3m = "alphapulldown.scripts.rename_colab_search_a3m:main"
prepare_seq_names = "alphapulldown.scripts.prepare_seq_names:main"
generate_crosslink_pickle = "alphapulldown.scripts.generate_crosslink_pickle:main"
convert_to_modelcif = "alphapulldown.scripts.entry_points:convert_to_modelcif"
run_structure_prediction = "alphapulldown.scripts.entry_points:run_structure_prediction"
